Renamed org.holodeckb2b.integ package into org.holodeckb2b.test.integration.
