+++ modules/holodeckb2b-it/pom.xml
@@ -109,7 +109,7 @@
                         <includes>
                           <include>**/*.java</include>
                         </includes>
-                        <groups>org.holodeckb2b.integ.IntegrationTest</groups>
+                        <groups>org.holodeckb2b.test.integration.IntegrationTest</groups>
                     </configuration>
                     <executions>
                         <execution>
