+++ modules/holodeckb2b-it/src/test/java/org/holodeckb2b/test/integration/OneWayMEPWorkflows.java
@@ -14,7 +14,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.holodeckb2b.integ;
+package org.holodeckb2b.test.integration;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
